LOK-2111: Make sure default policy have default tag when custom tag is empty (#1493)
* LOK-2111: Make sure default policy have default tag when custom tag is empty

* LOK-2111: handle update with only system tags

* LOK-2111: make system policy tag support null tag for empty tag support

+++ alert/src/main/java/org/opennms/horizon/alertservice/service/MonitorPolicyService.java
@@ -57,6 +57,7 @@ import javax.ws.rs.NotFoundException;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.function.Function;
@@ -122,7 +123,7 @@ public class MonitorPolicyService {
             SystemPolicyTag defaultPolicyTag;
             var existingTag = tagRepository.findByTenantIdAndName(tenantId, tag.getName());
             var matchedTag = existingTagMap.get(tag.getName());
-            if (matchedTag == null) {
+            if (matchedTag == null || !tenantId.equals(matchedTag.getTenantId())) {
                 Tag updatedTag;
                 if (existingTag.isEmpty()) {
                     tag.setPolicies(new HashSet<>());
@@ -142,11 +143,25 @@ public class MonitorPolicyService {
 
         var removedTags = new HashSet<>(Sets.difference(defaultPolicy.getTags(), newTags));
         removedTags.forEach(tag -> {
-                systemPolicyTagRepository.deleteById(new SystemPolicyTag.RelationshipId(tenantId, defaultPolicy.getId(), tag));
                 defaultPolicy.getTags().remove(tag);
+                if (SYSTEM_TENANT.equals(tag.getTenantId())) {
+                    return;
+                }
+                systemPolicyTagRepository.deleteById(new SystemPolicyTag.RelationshipId(tenantId, defaultPolicy.getId(), tag));
             }
         );
+        if (!newTags.isEmpty()) {
+            systemPolicyTagRepository.deleteEmptyTagByTenantIdAndPolicyId(tenantId, defaultPolicy.getId());
+        } else if (!removedTags.isEmpty()) {
+            var systemPolicyTag = new SystemPolicyTag(tenantId, defaultPolicy.getId(), null);
+            systemPolicyTagRepository.save(systemPolicyTag);
+        }
 
+        existingTags.forEach(t -> {
+            if (SYSTEM_TENANT.equals(t.getTenantId())) {
+                t.setTenantId(tenantId);
+            }
+        });
         handleTagOperationUpdate(existingTags, newTags);
 
         return policyMapper.map(defaultPolicy);
@@ -227,9 +242,16 @@ public class MonitorPolicyService {
     private Optional<MonitorPolicy> getDefaultPolicy(String tenantId) {
         return repository.findByNameAndTenantId(DEFAULT_POLICY, SYSTEM_TENANT)
             .map(p -> {
-                var tags = systemPolicyTagRepository.findByTenantIdAndPolicyId(tenantId, p.getId())
-                    .stream().map(SystemPolicyTag::getTag).collect(Collectors.toSet());
-                p.setTags(tags);
+                var systemPolicyTags = systemPolicyTagRepository.findByTenantIdAndPolicyId(tenantId, p.getId());
+                var tags = systemPolicyTags.stream().map(systemPolicyTag -> {
+                    if (systemPolicyTag == null) {
+                        return null;
+                    }
+                    return systemPolicyTag.getTag();
+                }).filter(Objects::nonNull).collect(Collectors.toSet());
+                if (!systemPolicyTags.isEmpty()) {
+                    p.setTags(tags);
+                }
                 return p;
             });
     }
